requirements.txt
numpy==2.0.2
Pillow==11.0.0
piexif==1.1.3
tensorflow==2.18.0
tensorflow_hub==0.16.1
torch==2.5.1+cu124
torchvision==0.20.1+cu124
scipy==1.14.1